leagueOfDrivers_BE/requirement.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==2.1.3
django-activity-stream==0.7.0
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-formtools==2.1
django-import-export==1.1.0
django-rest-framework-nested==0.0.1
django-rest-swagger==2.2.0
django-reversion==3.0.1
django-tagging==0.4.6
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.9.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.0
Naked==0.1.31
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.9
Pillow==5.3.0
psycopg2==2.7.5
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.7
PyYAML==3.13
requests==2.20.0
shellescape==3.4.1
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
